:truck: Move generic API tests to top-level tests
These tests essentially ensure that the correct middleware stack is
configured, and that the correct exception handler for API views is
defined.
