+++ src/openzaak/tests/__init__.py
@@ -1,0 +1,6 @@
+"""
+Execute general-purpose tests.
+
+The tests in this package are not tied to any specific component, but rather
+to the global project/API configuration.
+"""
