+++ src/openzaak/tests/api_strategy/__init__.py
@@ -1,0 +1,11 @@
+"""
+Test general API strategy guidelines.
+
+Test the `national API guidelines`_ and/or the `DSO API guidelines`_. Just for
+your information - the national guidelines are a result of the efforts done by
+the DSO to compose API guidelines.
+
+.. _national API guidelines: https://github.com/Geonovum/KP-APIs
+.. _DSO API guidelines: https://aandeslagmetdeomgevingswet.nl/digitaal-stelsel\
+  /technisch-aansluiten/standaarden/api-uri-strategie/
+"""
