+++ src/openzaak/tests/api_strategy/urls.py
@@ -1,3 +1,9 @@
+"""
+Define a dummy ROOT_URLCONF for testing.
+
+This enables us to test (a) view(s), processing the entire request/response
+processing stack, including middleware.
+"""
 from django.urls import path
 
 from rest_framework.response import Response
